@@ -1,51 +1,58 @@
 {
   "name": "personal-site",
-  "version": "3.0.1",
-  "homepage": "https://fmjohnson97.gitub.io",
+  "version": "4.0.0",
+  "homepage": "https://mldangelo.com/",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/fmjohnson97/fmjohnson97.gituhub.io"
+  },
+  "engines": {
+    "node": ">=14.x"
   },
   "scripts": {
     "build": "rimraf ./build && react-scripts build",
     "predeploy": "npm run build && react-snap",
     "deploy": "gh-pages -d build",
     "start": "react-scripts start",
-    "lint": "eslint src",
-    "test": "react-scripts test --watchAll=false",
+    "lint": "eslint $(git ls-files '*.js')",
+    "test": "npx jest",
     "analyze": "npm run build && source-map-explorer build/static/js/*.chunk.js"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-brands-svg-icons": "^5.15.2",
-    "@fortawesome/free-regular-svg-icons": "^5.15.2",
-    "@fortawesome/react-fontawesome": "0.1.14",
-    "dayjs": "^1.10.4",
-    "gh-pages": "^3.1.0",
-    "node-sass": "^4.14.1",
-    "prop-types": "^15.7.2",
-    "raw.macro": "^0.4.2",
-    "react": "^17.0.1",
-    "react-burger-menu": "^3.0.3",
-    "react-dom": "^17.0.1",
-    "react-ga": "^3.3.0",
-    "react-helmet-async": "^1.0.7",
-    "react-markdown": "^5.0.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
-    "react-snap": "^1.23.0",
-    "rimraf": "^3.0.2"
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-brands-svg-icons": "^6.4.2",
+    "@fortawesome/free-regular-svg-icons": "^6.4.2",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "dayjs": "^1.11.9",
+    "gh-pages": "^5.0.0",
+    "markdown-to-jsx": "^7.3.2",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-burger-menu": "^3.0.9",
+    "react-dom": "^18.2.0",
+    "react-ga": "^3.3.1",
+    "react-helmet-async": "^1.3.0",
+    "react-router-dom": "^6.15.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.3",
-    "eslint": "^7.19.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "^7.22.0",
-    "source-map-explorer": "^2.5.2"
+    "react-snap": "^1.23.0",
+    "rimraf": "^5.0.1",
+    "babel-jest": "^29.6.4",
+    "@babel/eslint-parser": "^7.22.15",
+    "@babel/preset-env": "^7.22.10",
+    "@babel/preset-react": "^7.22.5",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "eslint": "^8.49.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-react": "^7.33.2",
+    "sass": "^1.66.1",
+    "source-map-explorer": "^2.5.3"
   },
   "browserslist": {
     "production": [
