@@ -1,30 +1,32 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 
 import CategoryButton from './Skills/CategoryButton';
 import SkillBar from './Skills/SkillBar';
 
-const handleProps = ({ categories, skills }) => ({
-  buttons: categories.map((cat) => cat.name).reduce((obj, key) => ({
-    ...obj,
-    [key]: false,
-  }), { All: true }),
-  skills,
-});
+const Skills = ({ skills, categories }) => {
+  const initialButtons = Object.fromEntries([['All', false]].concat(categories.map(({ name }) => [name, false])));
 
-class Skills extends Component {
-  constructor(props) {
-    super(props);
-    this.state = handleProps({ categories: props.categories, skills: props.skills });
-  }
+  const [buttons, setButtons] = useState(initialButtons);
 
-  getRows() {
+  const handleChildClick = (label) => {
+    // Toggle button that was clicked. Turn all other buttons off.
+    const newButtons = Object.keys(buttons).reduce((obj, key) => ({
+      ...obj,
+      [key]: (label === key) && !buttons[key],
+    }), {});
+    // Turn on 'All' button if other buttons are off
+    newButtons.All = !Object.keys(buttons).some((key) => newButtons[key]);
+    setButtons(newButtons);
+  };
+
+  const getRows = () => {
     // search for true active categories
-    const actCat = Object.keys(this.state.buttons).reduce((cat, key) => (
-      this.state.buttons[key] ? key : cat
+    const actCat = Object.keys(buttons).reduce((cat, key) => (
+      buttons[key] ? key : cat
     ), 'All');
 
-    return this.state.skills.sort((a, b) => {
+    const comparator = (a, b) => {
       let ret = 0;
       if (a.competency > b.competency) ret = -1;
       else if (a.competency < b.competency) ret = 1;
@@ -33,60 +35,45 @@ class Skills extends Component {
       else if (a.title > b.title) ret = 1;
       else if (a.title < b.title) ret = -1;
       return ret;
-    }).filter((skill) => (actCat === 'All' || skill.category.includes(actCat)))
+    };
+
+    return skills.sort(comparator).filter((skill) => (actCat === 'All' || skill.category.includes(actCat)))
       .map((skill) => (
         <SkillBar
-          categories={this.props.categories}
+          categories={categories}
           data={skill}
           key={skill.title}
         />
       ));
-  }
+  };
 
-  getButtons() {
-    return Object.keys(this.state.buttons).map((key) => (
-      <CategoryButton
-        label={key}
-        key={key}
-        active={this.state.buttons}
-        handleClick={this.handleChildClick}
-      />
-    ));
-  }
+  const getButtons = () => Object.keys(buttons).map((key) => (
+    <CategoryButton
+      label={key}
+      key={key}
+      active={buttons}
+      handleClick={handleChildClick}
+    />
+  ));
 
-  handleChildClick = (label) => {
-    this.setState((prevState) => {
-      // Toggle button that was clicked. Turn all other buttons off.
-      const buttons = Object.keys(prevState.buttons).reduce((obj, key) => ({
-        ...obj,
-        [key]: (label === key) && !prevState.buttons[key],
-      }), {});
-      // Turn on 'All' button if other buttons are off
-      buttons.All = !Object.keys(prevState.buttons).some((key) => buttons[key]);
-      return { buttons };
-    });
-  }
-
-  render() {
-    return (
-      <div className="skills">
-        <div className="link-to" id="skills" />
-        <div className="title">
-          <h3>Skills</h3>
-          <p>Note: I think these sections are silly, but everyone seems to have one.
-            Here is a *mostly* honest overview of my skills.
-          </p>
-        </div>
-        <div className="skill-button-container">
-          {this.getButtons()}
-        </div>
-        <div className="skill-row-container">
-          {this.getRows()}
-        </div>
+  return (
+    <div className="skills">
+      <div className="link-to" id="skills" />
+      <div className="title">
+        <h3>Skills</h3>
+        <p>Note: I think these sections are silly, but everyone seems to have one.
+          Here is a *mostly* honest overview of my skills.
+        </p>
       </div>
-    );
-  }
-}
+      <div className="skill-button-container">
+        {getButtons()}
+      </div>
+      <div className="skill-row-container">
+        {getRows()}
+      </div>
+    </div>
+  );
+};
 
 Skills.propTypes = {
   skills: PropTypes.arrayOf(PropTypes.shape({
