@@ -1,4 +1,5 @@
 import React, { useState, useEffect, useRef } from 'react';
+import PropTypes from 'prop-types';
 
 // Validates the first half of an email address.
 const validateText = (text) => {
@@ -44,13 +45,13 @@ const useInterval = (callback, delay) => {
   }, [delay]);
 };
 
-const EmailLink = () => {
+const EmailLink = ({ loopMessage }) => {
   const hold = 50; // ticks to wait after message is complete before rendering next message
   const delay = 50; // tick length in mS
 
   const [idx, updateIter] = useState(0); // points to current message
   const [message, updateMessage] = useState(messages[idx]);
-  const [char, updateChar] = useState(messages[idx].length); // points to current char
+  const [char, updateChar] = useState(0); // points to current char
   const [isActive, setIsActive] = useState(true); // disable when all messages are printed
 
   useInterval(() => {
@@ -61,7 +62,12 @@ const EmailLink = () => {
       newChar = 0;
     }
     if (newIdx === messages.length) {
-      setIsActive(false);
+      if (loopMessage) {
+        updateIter(0);
+        updateChar(0);
+      } else {
+        setIsActive(false);
+      }
     } else {
       updateMessage(messages[newIdx].slice(0, newChar));
       updateIter(newIdx);
@@ -82,6 +88,14 @@ const EmailLink = () => {
       </a>
     </div>
   );
+};
+
+EmailLink.defaultProps = {
+  loopMessage: false,
+};
+
+EmailLink.propTypes = {
+  loopMessage: PropTypes.bool,
 };
 
 export default EmailLink;
