@@ -139,7 +139,7 @@ const skills = [
   },
 ].map((skill) => ({ ...skill, category: skill.category.sort() }));
 
-// this is a list of colors that I like. The length should be == to the
+// this is a list of colors that I like. The length should be === to the
 // number of categories. Re-arrange this list until you find a pattern you like.
 const colors = [
   '#6968b3',
@@ -156,12 +156,10 @@ const colors = [
 ];
 
 const categories = [
-  ...new Set(skills.reduce((acc, { category }) => acc.concat(category), [])),
-]
-  .sort()
-  .map((category, index) => ({
-    name: category,
-    color: colors[index],
-  }));
+  ...new Set(skills.flatMap(({ category }) => category)),
+].sort().map((category, index) => ({
+  name: category,
+  color: colors[index],
+}));
 
 export { categories, skills };
