@@ -1,5 +1,5 @@
 import React from 'react';
-import { hydrate, render } from 'react-dom';
+import { createRoot, hydrateRoot } from 'react-dom/client';
 import App from './App';
 
 // See https://reactjs.org/docs/strict-mode.html
@@ -13,7 +13,8 @@ const rootElement = document.getElementById('root');
 
 // hydrate is required by react-snap.
 if (rootElement.hasChildNodes()) {
-  hydrate(<StrictApp />, rootElement);
+  hydrateRoot(rootElement, <StrictApp />);
 } else {
-  render(<StrictApp />, rootElement);
+  const root = createRoot(rootElement);
+  root.render(<StrictApp />);
 }
