@@ -8,7 +8,7 @@ const PageNotFound = () => (
       <Helmet title="404 Not Found">
         <meta name="description" content="The content you are looking for cannot be found." />
       </Helmet>
-      <h1 data-testid="heading">Page Not Found</h1>
+      <h1>Page Not Found</h1>
       <p>Return <Link to="/">home</Link>.</p>
     </div>
   </HelmetProvider>
