@@ -33,23 +33,18 @@ const Resume = () => (
     <article className="post" id="resume">
       <header>
         <div className="title">
-          <h2 data-testid="heading"><Link to="resume">Resume</Link></h2>
+          <h2><Link to="resume">Resume</Link></h2>
           <div className="link-container">
-            {sections.map((sec) => (
+            {Object.keys(sections).map((sec) => (
               <h4 key={sec}>
                 <a href={`#${sec.toLowerCase()}`}>{sec}</a>
               </h4>))}
           </div>
-
         </div>
       </header>
-      <Education data={degrees} />
-      <Experience data={positions} />
-      <Leadership data={leadership} />
-      <Skills skills={skills} categories={categories} />
-      <Courses data={courses} />
-      <References />
-
+      {Object.entries(sections).map(([name, Section]) => (
+        <Section key={name} />
+      ))}
     </article>
   </Main>
 );
